slider synchronization

+++ app/src/main/java/com/example/lyricsgame/ui/game/GameUiState.kt
@@ -2,4 +2,10 @@ package com.example.lyricsgame.ui.game
 
 import com.example.lyricsgame.data.model.Track
 
-data class GameUiState(var remainingTimeToStartGame: Int = 3, var trackList: List<Track>? = listOf(), var currentPosition: Int = 0, var currentTrack: Track? = null, var sliderPosition: Float = 0F)
+data class GameUiState(
+    var remainingTimeToStartGame: Int = 3,
+    var trackList: List<Track>? = listOf(),
+    var currentPosition: Int = 0,
+    var currentTrack: Track? = null,
+    var sliderPosition: Int = 0
+)
